cleanup delete accounts endpoint

+++ SafeAccountsAPI/Controllers/UsersController.cs
@@ -370,11 +370,11 @@ namespace SafeAccountsAPI.Controllers
                 // if this user does not own the folder we are adding to, then error
                 if (accToAdd.FolderID != null && !_context.Users.Single(a => a.ID == id).Folders.Exists(b => b.ID == accToAdd.FolderID))
                 {
-                    ErrorMessage error = new ErrorMessage("Failed to create new account", "User does not own the folder they wish to add an account to.");
+                    ErrorMessage error = new ErrorMessage("Failed to create new account", "User does not have a folder matching that ID.");
                     return new BadRequestObjectResult(error);
                 }
 
-                // use token in header to to 
+                // create new account and save it
                 Account new_account = new Account(accToAdd, id);
                 _context.Accounts.Add(new_account);
                 _context.SaveChanges();
@@ -388,27 +388,33 @@ namespace SafeAccountsAPI.Controllers
         }
 
         [HttpDelete("{id:int}/accounts/{account_id:int}")] // working
-        public string User_DeleteAccount(int id, int account_id)
+        public IActionResult User_DeleteAccount(int id, int account_id)
         {
             // verify that the user is either admin or is requesting their own data
             if (!HelperMethods.ValidateIsUserOrAdmin(_httpContextAccessor, _context, id))
             {
-                Response.StatusCode = 401;
-                return JObject.FromObject(new ErrorMessage("Invalid User", "Caller can only access their information.")).ToString();
+                ErrorMessage error = new ErrorMessage("Invalid User", "Caller can only access their information.");
+                return new UnauthorizedObjectResult(error);
             }
 
             try
             {
+                // validate ownership of said account
+                if (!_context.Users.Single(a => a.ID == id).Accounts.Exists(b => b.ID == account_id))
+                {
+                    ErrorMessage error = new ErrorMessage("Failed to delete account", "User does not have an account matching that ID.");
+                    return new BadRequestObjectResult(error);
+                }
+
                 _context.Accounts.Remove(_context.Users.Single(a => a.ID == id).Accounts.Single(b => b.ID == account_id)); // fist match user id to ensure ownership
                 _context.SaveChanges();
+                return Ok();
             }
             catch (Exception ex)
             {
-                Response.StatusCode = 500;
-                return JObject.FromObject(new ErrorMessage("Error deleting account.", ex.Message)).ToString();
+                ErrorMessage error = new ErrorMessage("Error deleting account.", ex.Message);
+                return new InternalServerErrorResult(error);
             }
-
-            return SuccessMessage.Result;
         }
 
         // get a specific accounts info
